Acceso_a_Datos/PR5 - Asistente virtual/Asistente_miriam/requirements.txt
google-genai
python-dotenv



